chore: v2 API를 패키지 단위로 완전히 분리
