+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v1/dto/request/SubmitWorkbookRequest.java
@@ -1,4 +1,4 @@
-package com.umc.product.curriculum.adapter.in.web.dto.request;
+package com.umc.product.curriculum.adapter.in.web.v1.dto.request;
 
 import com.umc.product.curriculum.application.port.in.command.SubmitWorkbookCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
