+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v1/swagger/AdminCurriculumControllerApi.java
@@ -1,6 +1,6 @@
-package com.umc.product.curriculum.adapter.in.web.swagger;
+package com.umc.product.curriculum.adapter.in.web.v1.swagger;
 
-import com.umc.product.curriculum.adapter.in.web.dto.request.ManageCurriculumRequest;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.request.ManageCurriculumRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
