+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v1/swagger/WorkbookQueryControllerApi.java
@@ -1,9 +1,9 @@
-package com.umc.product.curriculum.adapter.in.web.swagger;
+package com.umc.product.curriculum.adapter.in.web.v1.swagger;
 
 import com.umc.product.common.domain.enums.ChallengerPart;
-import com.umc.product.curriculum.adapter.in.web.dto.response.AvailableWeeksResponse;
-import com.umc.product.curriculum.adapter.in.web.dto.response.StudyGroupFilterResponse;
-import com.umc.product.curriculum.adapter.in.web.dto.response.WorkbookSubmissionResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.AvailableWeeksResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.StudyGroupFilterResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.WorkbookSubmissionResponse;
 import com.umc.product.global.response.CursorResponse;
 import com.umc.product.global.security.MemberPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
