+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v2/CurriculumQueryV2Controller.java
@@ -1,9 +1,9 @@
 package com.umc.product.curriculum.adapter.in.web.v2;
 
 import com.umc.product.common.domain.enums.ChallengerPart;
-import com.umc.product.curriculum.adapter.in.web.dto.response.CurriculumProgressResponse;
-import com.umc.product.curriculum.adapter.in.web.dto.response.CurriculumResponse;
-import com.umc.product.curriculum.adapter.in.web.swagger.v2.CurriculumQueryV2ControllerApi;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.CurriculumProgressResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.CurriculumResponse;
+import com.umc.product.curriculum.adapter.in.web.v2.swagger.CurriculumQueryV2ControllerApi;
 import com.umc.product.curriculum.application.port.in.query.GetCurriculumProgressUseCase;
 import com.umc.product.curriculum.application.port.in.query.GetCurriculumUseCase;
 import com.umc.product.global.security.MemberPrincipal;
