+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v2/swagger/CurriculumQueryV2ControllerApi.java
@@ -1,11 +1,13 @@
-package com.umc.product.curriculum.adapter.in.web.swagger.v2;
+package com.umc.product.curriculum.adapter.in.web.v2.swagger;
 
 import com.umc.product.common.domain.enums.ChallengerPart;
-import com.umc.product.curriculum.adapter.in.web.dto.response.CurriculumProgressResponse;
-import com.umc.product.curriculum.adapter.in.web.dto.response.CurriculumResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.CurriculumProgressResponse;
+import com.umc.product.curriculum.adapter.in.web.v1.dto.response.CurriculumResponse;
 import com.umc.product.global.security.MemberPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
@@ -18,13 +20,13 @@ public interface CurriculumQueryV2ControllerApi {
             "`week` 파라미터를 지정하면 해당 주차의 워크북만 반환합니다."
     )
     @ApiResponses(value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
+        @ApiResponse(
             responseCode = "200",
             description = "조회 성공"
         )
     })
     CurriculumResponse getCurriculum(
-        @Parameter(description = "기수 ID", required = true, in = io.swagger.v3.oas.annotations.enums.ParameterIn.PATH) Long gisuId,
+        @Parameter(description = "기수 ID", required = true, in = ParameterIn.PATH) Long gisuId,
         @Parameter(description = "파트", required = true) ChallengerPart part,
         @Parameter(description = "주차 번호 (생략 시 전체 주차 반환)") Integer week
     );
@@ -35,13 +37,13 @@ public interface CurriculumQueryV2ControllerApi {
             "각 주차별 워크북의 상태(기본/진행중/제출완료/통과/실패)를 반환합니다."
     )
     @ApiResponses(value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
+        @ApiResponse(
             responseCode = "200",
             description = "조회 성공"
         )
     })
     CurriculumProgressResponse getMyProgress(
-        @Parameter(description = "기수 ID", required = true, in = io.swagger.v3.oas.annotations.enums.ParameterIn.PATH) Long gisuId,
+        @Parameter(description = "기수 ID", required = true, in = ParameterIn.PATH) Long gisuId,
         @Parameter(hidden = true) MemberPrincipal memberPrincipal
     );
 }
